requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.3
django-ninja==1.3.0
gunicorn==23.0.0
packaging==24.2
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
PyMySQL==1.1.1
sqlparse==0.5.2
typing_extensions==4.12.2
